+++ changelog.md
@@ -1,3 +1,4 @@
 - fix crash when negating rune of spirit tools sacrifice
 - fix division by zero in power distribution
 - added/fixed star indicator for power distribution
+- fixed rune of heights offset